#182 refactor the ReflectionUtils class: Exctracted view loading specific methods into its own util class
