+++ mvvmfx/src/test/java/de/saxsys/mvvmfx/internal/viewloader/ViewLoaderReflectionUtilsTest.java
@@ -7,7 +7,7 @@ import org.junit.Test;
 import de.saxsys.mvvmfx.ViewModel;
 import de.saxsys.mvvmfx.internal.viewloader.example.TestViewModel;
 
-public class ReflectionUtilsTest {
+public class ViewLoaderReflectionUtilsTest {
 	
 	
 	@Test
@@ -15,7 +15,7 @@ public class ReflectionUtilsTest {
 		class TestView implements View<TestViewModel> {
 		}
 
-		ViewModel viewModel = ReflectionUtils.createViewModel(new TestView());
+		ViewModel viewModel = ViewLoaderReflectionUtils.createViewModel(new TestView());
 		
 		assertThat(viewModel).isNotNull().isInstanceOf(TestViewModel.class);
 	}
@@ -25,7 +25,7 @@ public class ReflectionUtilsTest {
 		class TestView implements View{
 		}
 
-		ViewModel viewModel = ReflectionUtils.createViewModel(new TestView());
+		ViewModel viewModel = ViewLoaderReflectionUtils.createViewModel(new TestView());
 		
 		assertThat(viewModel).isNull();
 	}
@@ -35,7 +35,7 @@ public class ReflectionUtilsTest {
 		class TestView implements View<ViewModel>{
 		}
 		
-		ViewModel viewModel = ReflectionUtils.createViewModel(new TestView());
+		ViewModel viewModel = ViewLoaderReflectionUtils.createViewModel(new TestView());
 		
 		assertThat(viewModel).isNull();
 	}
